remove dockerpy wrappers
