+++ neurodocker/__init__.py
@@ -1,13 +1,8 @@
-# Author: Jakub Kaczmarzyk <jakubk@mit.edu>
-
-from __future__ import absolute_import, unicode_literals
-
 import sys
 
 from neurodocker.version import __version__
-
-from neurodocker.docker import DockerContainer, DockerImage
-from neurodocker.generators import Dockerfile, SingularityRecipe
+from neurodocker.generators import Dockerfile
+from neurodocker.generators import SingularityRecipe
 from neurodocker.utils import set_log_level
 
 if sys.version_info[0] < 3:
