--- neurodocker/docker/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import absolute_import
-
-from neurodocker.docker.docker import (client, copy_file_to_container,
-                                       copy_file_from_container,
-                                       DockerContainer, DockerImage,)
